Add a unique ID system for all new jots.
These changes add the ability to associate a unique ID with each jot.
This is a new way the user can reference a particular jot that won't
change when a jot earlier in the list gets deleted.

Right now we randomly generate a string that is unique for you.
Eventually we could let users change/set these fields to create named
jots.

These changes are backwards compatible with jots that don't have these
metadata, in this case we will just use the old spot within the list.